+++ src/query/range_query/mod.rs
@@ -2,10 +2,12 @@ use crate::schema::Type;
 
 mod fast_field_range_doc_set;
 mod range_query;
-mod range_query_u64_fastfield;
+mod range_query_fastfield;
+
+pub use common::bounds::BoundsRange;
 
 pub use self::range_query::*;
-pub use self::range_query_u64_fastfield::FastFieldRangeWeight;
+pub use self::range_query_fastfield::*;
 
 // TODO is this correct?
 pub(crate) fn is_type_valid_for_fastfield_range_query(typ: Type) -> bool {
